neuroshorts_automation/requirements.txt
openai>=1.68.2
python-dotenv>=1.0.1
moviepy>=2.1.1
Pillow>=10.4.0
numpy>=1.26.4,<2.1
